refactor : 프로젝트 등록 페이지 최적화 및 폼 유효성 검사 추가 - 인라인 스타일 테일윈드 클래스로 변경 -ssr 최적화를 위해 클라이언트 로직을 별도 컴포넌트로 분리
